backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai>=1.0.0
sentence-transformers==2.3.1
faiss-cpu==1.9.0
rank-bm25==0.2.2
sqlalchemy==2.0.25
pydantic-settings==2.1.0
python-dotenv==1.0.0
numpy==1.26.3
tiktoken==0.5.2
pytest==7.4.4
httpx==0.26.0
python-multipart==0.0.6
